changed doubles to floats, and initial MPI setup and hello world in place

+++ lbm.h
@@ -13,9 +13,9 @@ typedef struct {
     int ny;            /* no. of cells in y-direction */
     int max_iters;      /* no. of iterations */
     int reynolds_dim;  /* dimension for Reynolds number */
-    double density;       /* density per link */
-    double accel;         /* density redistribution */
-    double omega;         /* relaxation parameter */
+    float density;       /* density per link */
+    float accel;         /* density redistribution */
+    float omega;         /* relaxation parameter */
 } param_t;
 
 /* obstacle positions */
@@ -28,7 +28,7 @@ typedef struct {
 
 /* struct to hold the 'speed' values */
 typedef struct {
-    double speeds[NSPEEDS];
+    float speeds[NSPEEDS];
 } speed_t;
 
 typedef enum { ACCEL_ROW, ACCEL_COLUMN } accel_e;
@@ -43,13 +43,13 @@ void parse_args (int argc, char* argv[],
 
 void initialise(const char* paramfile, accel_area_t * accel_area,
     param_t* params, speed_t** cells_ptr, speed_t** tmp_cells_ptr,
-    int** obstacles_ptr, double** av_vels_ptr);
+    int** obstacles_ptr, float** av_vels_ptr);
 
 void write_values(const char * final_state_file, const char * av_vels_file,
-    const param_t params, speed_t* cells, int* obstacles, double* av_vels);
+    const param_t params, speed_t* cells, int* obstacles, float* av_vels);
 
 void finalise(speed_t** cells_ptr, speed_t** tmp_cells_ptr,
-    int** obstacles_ptr, double** av_vels_ptr);
+    int** obstacles_ptr, float** av_vels_ptr);
 
 void timestep(const param_t params, const accel_area_t accel_area,
     speed_t* cells, speed_t* tmp_cells, int* obstacles);
@@ -61,13 +61,13 @@ void collision(const param_t params, speed_t* cells, speed_t* tmp_cells, int* ob
 
 /* Sum all the densities in the grid.
 ** The total should remain constant from one timestep to the next. */
-double total_density(const param_t params, speed_t* cells);
+float total_density(const param_t params, speed_t* cells);
 
 /* compute average velocity */
-double av_velocity(const param_t params, speed_t* cells, int* obstacles);
+float av_velocity(const param_t params, speed_t* cells, int* obstacles);
 
 /* calculate Reynolds number */
-double calc_reynolds(const param_t params, speed_t* cells, int* obstacles);
+float calc_reynolds(const param_t params, speed_t* cells, int* obstacles);
 
 /* Exit, printing out formatted string */
 #define DIE(...) exit_with_error(__LINE__, __FILE__, __VA_ARGS__)
