+++ utils.c
@@ -96,12 +96,12 @@ void parse_args (int argc, char* argv[],
 
 void initialise(const char* param_file, accel_area_t * accel_area,
     param_t* params, speed_t** cells_ptr, speed_t** tmp_cells_ptr,
-    int** obstacles_ptr, double** av_vels_ptr)
+    int** obstacles_ptr, float** av_vels_ptr)
 {
     FILE   *fp;            /* file pointer */
     int    ii,jj, kk;          /* generic counters */
     int    retval;         /* to hold return value for checking */
-    double w0,w1,w2;       /* weighting factors */
+    float w0,w1,w2;       /* weighting factors */
 
     /* Rectangular obstacles */
     int n_obstacles;
@@ -123,11 +123,11 @@ void initialise(const char* param_file, accel_area_t * accel_area,
     if (retval != 1) DIE("Could not read param file: max_iters");
     retval = fscanf(fp,"%d\n",&(params->reynolds_dim));
     if (retval != 1) DIE("Could not read param file: reynolds_dim");
-    retval = fscanf(fp,"%lf\n",&(params->density));
+    retval = fscanf(fp,"%f\n",&(params->density));
     if (retval != 1) DIE("Could not read param file: density");
-    retval = fscanf(fp,"%lf\n",&(params->accel));
+    retval = fscanf(fp,"%f\n",&(params->accel));
     if (retval != 1) DIE("Could not read param file: accel");
-    retval = fscanf(fp,"%lf\n",&(params->omega));
+    retval = fscanf(fp,"%f\n",&(params->omega));
     if (retval != 1) DIE("Could not read param file: omega");
 
     if (params->nx < 100) DIE("x dimension of grid in input file was too small (must be >100)");
@@ -188,7 +188,7 @@ void initialise(const char* param_file, accel_area_t * accel_area,
     *obstacles_ptr = (int*) malloc(sizeof(int)*(params->ny*params->nx));
     if (*obstacles_ptr == NULL) DIE("Cannot allocate memory for patches");
 
-    *av_vels_ptr = (double*) malloc(sizeof(double)*(params->max_iters));
+    *av_vels_ptr = (float*) malloc(sizeof(float)*(params->max_iters));
     if (*av_vels_ptr == NULL) DIE("Cannot allocate memory for av_vels");
 
     w0 = params->density * 4.0/9.0;
@@ -243,7 +243,7 @@ void initialise(const char* param_file, accel_area_t * accel_area,
 }
 
 void finalise(speed_t** cells_ptr, speed_t** tmp_cells_ptr,
-    int** obstacles_ptr, double** av_vels_ptr)
+    int** obstacles_ptr, float** av_vels_ptr)
 {
     /* Free allocated memory */
     free(*cells_ptr);
@@ -251,4 +251,3 @@ void finalise(speed_t** cells_ptr, speed_t** tmp_cells_ptr,
     free(*obstacles_ptr);
     free(*av_vels_ptr);
 }
-
